rename files & fix ci
